refactor(cli): move e2e module cluster into src/cli/e2e and rewire command
Relocate the e2e-only modules and their tests (clean-environment, grafana-health, exit-codes, playwright-runner, e2e-runner-contract, e2e-reporter, e2e-targets, e2e-package, recommender-resolver, manifest-preflight, guide-chains) from src/cli/utils into a dedicated src/cli/e2e feature directory. utils stays the shared layer (file-loader, graph-cycles, output) that e2e imports into. Rewire commands/e2e.ts and the tests/e2e-runner contract/reporter imports, colocate the manifest-preflight test, and fix the design-doc path reference.
