+++ src/cli/e2e/clean-environment.ts
