+++ src/cli/e2e/manifest-preflight.ts
